backend/rag_utils.py
import os
from openai import OpenAI
from qdrant_client import QdrantClient
from dotenv import load_dotenv

load_dotenv()

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
qdrant = QdrantClient(
    url=os.getenv("QDRANT_URL"),
    api_key=os.getenv("QDRANT_API_KEY")
)

def get_embedding(text):
    response = client.embeddings.create(
        input=text,
        model="text-embedding-3-small"
    )
    return response.data[0].embedding

def search_content(query, limit=3):
    query_vector = get_embedding(query)
    search_result = qdrant.search(
        collection_name="textbook_content",
        query_vector=query_vector,
        limit=limit
    )
    return search_result
