backend/requirements.txt
fastapi
uvicorn
openai
qdrant-client
psycopg2-binary
python-dotenv
pydantic
